Make the polling delay dynamic
Since the previous implementation uses a single poll-based task with a
fixed sleep time, we could run into the situation of a poll inside
`debounce` happening just before another invocation of the event hook.
In that case, we would wait for the full delay time, poll again, notice
that the handler has been called and enter another iteration of
sleeping and polling.

We can circumvent this by recording a timestamp of the last handler
invocation and then calculating the next poll to happen right after the
specified delay, relative to the latest invocation.