Make the polling delay dynamic (#2)
* Make the polling delay dynamic

Since the previous implementation uses a single poll-based task with a
fixed sleep time, we could run into the situation of a poll inside
`debounce` happening just before another invocation of the event hook.
In that case, we would wait for the full delay time, poll again, notice
that the handler has been called and enter another iteration of
sleeping and polling.

We can circumvent this by recording a timestamp of the last handler
invocation and then calculating the next poll to happen right after the
specified delay, relative to the latest invocation.

* Changes and Improvements

1. directly import `time.monolitic as now` as shorter alternative to creating
   intermediate `now` variables in debouncer.

2. reorder `time_to_wait` expression to start with static delay followed by
   dynamic part (diff between event trigger and now())

3. call `del call_at` in debounce() directly after checking escape condition
   to group `call_at` access related expressions

4. drop intermediate `view.view_id` in debounce() as it is called way less
   frequently then in wrapper() and thus isn't probably worth it.

5. drop intermediate `pending` variable

Step 3 and 5 prepare for better thread safety measures.

---------

Co-authored-by: deathaxe <deathaxe82@gmail.com>